pyproject.toml
[project]
name = "learn-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "firecrawl-py>=2.4.0",
    "langchain-aws>=0.2.22",
    "langchain-community>=0.3.22",
    "langchain-tavily>=0.1.6",
    "langchain-text-splitters>=0.3.8",
    "langchain[openai]>=0.3.24",
    "langgraph>=0.4.0",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "slack-sdk>=3.35.0",
    "tavily-python>=0.7.0",
    "wikipedia>=1.4.0",
    "youtube-search>=2.1.2",
]
